Add more security

+++ mercado/app/Http/Controllers/PostsController.php
@@ -80,6 +80,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post =  Post::find($id);
+
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -113,6 +118,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'unauthorized Page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
